Actualizado para Odoo v10.0 Eliminados archivos innecesarios
