+++ __init__.py
@@ -1,3 +1,2 @@
-# -*- coding: utf-8 -*-
-
+from . import models
 from . import models
