+++ security/sale_order_discount.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- Copyright 2017 Faros Inversiones Ltda.
+     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
+
+<odoo>
+<!--
+    <record model="ir.model.access" id="sale_order_discount_access_name">
+        <field name="name">sale_order_discount access name</field>
+        <field name="model_id" ref="model_sale_order_discount"/>
+
+        <field name="group_id" ref="base.group_user"/>
+        <field name="perm_read" eval="1"/>
+        <field name="perm_create" eval="0"/>
+        <field name="perm_write" eval="0"/>
+        <field name="perm_unlink" eval="0"/>
+    </record>
+-->
+</odoo>
