+++ views/invoice_global_discount.xml
@@ -1,12 +1,15 @@
-<openerp>
-  <data>
-<!-- 
+<?xml version="1.0" encoding="utf-8"?>
+<!-- Copyright 2017 Faros Inversiones Ltda.
+     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
+
+<odoo>
+
     <record model="ir.ui.view" id="sale_global_discount_invoice">
         <field name="name">sale_global_discount_invoice.form</field>
         <field name="model">account.invoice</field>
         <field name="inherit_id" ref="account.invoice_form"/>
         <field name="arch" type="xml">
-                <xpath expr="//field[@name='comment']" position="before">
+                <xpath expr="//field[@name='invoice_line_ids']" position="after">
                         <group name="discount" col="2" colspan="2">
                             <field name="discount_type" onchange="compute_discount"/>
                             <field name="discount_rate" onchange="compute_discount"/>
@@ -17,6 +20,5 @@
                 </xpath>
         </field>
     </record>
- -->
-  </data>
-</openerp>
+
+</odoo>
