+++ views/sale_order_discount.xml
@@ -1,0 +1,49 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- Copyright 2017 Faros Inversiones Ltda.
+     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
+
+<odoo>
+    <record model="ir.ui.view" id="sale_global_discount_sale_order">
+        <field name="name">sale_global_discount.form</field>
+        <field name="model">sale.order</field>
+        <field name="inherit_id" ref="sale.view_order_form"/>
+        <field name="arch" type="xml">
+                <xpath expr="//group[@name='sale_total']" position="before">
+                        <group name="discount" class="oe_left" colspan="2" col="2">
+                            <field name="discount_type" onchange="compute_discount"/>
+                            <field name="discount_rate" onchange="compute_discount"/>
+                        </group>
+                </xpath>
+                <xpath expr="//field[@name='amount_tax']" position="before">
+                            <field name="amount_discount" widget='monetary' options="{'currency_field': 'currency_id'}"/>
+                </xpath>
+        </field>
+    </record>
+
+<!--
+    <record model="ir.ui.view" id="sale_order_discount_form_view">
+        <field name="name">sale_order_discount.form (in sale_global_discount)</field>
+        <field name="model">sale_order_discount</field>
+        <field name="inherit_id" ref="TODO othermodule.form_view"/>
+        <field name="arch" type="xml">
+
+        </field>
+    </record>
+
+    <record model="ir.ui.view" id="sale_order_discount_search_view">
+        <field name="name">sale_order_discount.search (in sale_global_discount)</field>
+        <field name="model">sale_order_discount</field>
+        <field name="inherit_id" ref="TODO othermodule.search_view"/>
+        <field name="arch" type="xml">
+        </field>
+    </record>
+
+    <record model="ir.ui.view" id="sale_order_discount_tree_view">
+        <field name="name">sale_order_discount.tree (in sale_global_discount)</field>
+        <field name="model">sale_order_discount</field>
+        <field name="inherit_id" ref="TODO othermodule.tree_view"/>
+        <field name="arch" type="xml">
+        </field>
+    </record>
+-->
+</odoo>
