--- views/views.xml
@@ -1,39 +1,0 @@
-<openerp>
-  <data>
-
-    <record model="ir.ui.view" id="sale_global_discount_sale_order">
-        <field name="name">sale_global_discount.form</field>
-        <field name="model">sale.order</field>
-        <field name="inherit_id" ref="sale.view_order_form"/>
-        <field name="arch" type="xml">
-                <xpath expr="//group[@name='sale_total']" position="before">
-                        <group name="discount" class="oe_left" colspan="2" col="2">
-                            <field name="discount_type" onchange="compute_discount"/>
-                            <field name="discount_rate" onchange="compute_discount"/>
-                        </group>
-                </xpath>
-                <xpath expr="//field[@name='amount_tax']" position="before">
-                            <field name="amount_discount" widget='monetary' options="{'currency_field': 'currency_id'}"/>
-                </xpath>
-        </field>
-    </record>
-
-    <record model="ir.ui.view" id="sale_global_discount_invoice">
-        <field name="name">sale_global_discount_invoice.form</field>
-        <field name="model">account.invoice</field>
-        <field name="inherit_id" ref="account.invoice_form"/>
-        <field name="arch" type="xml">
-                <xpath expr="//field[@name='invoice_line_ids']" position="after">
-                        <group name="discount" col="2" colspan="2">
-                            <field name="discount_type" onchange="compute_discount"/>
-                            <field name="discount_rate" onchange="compute_discount"/>
-                        </group>
-                </xpath>
-                <xpath expr="//field[@name='amount_tax']" position="before">
-                            <field name="amount_discount" widget='monetary' options="{'currency_field': 'currency_id'}"/>
-                </xpath>
-        </field>
-    </record>
-
-  </data>
-</openerp>
